mm/gup: introduce pin_user_pages_fd() for pinning shmem/hugetlbfs file pages
Patch series "mm/gup: Introduce pin_user_pages_fd() for pinning
shmem/hugetlbfs file pages".

The first patch introduces pin_user_pages_fd() API and the second patch
shows how the udmabuf driver can use it to longterm-pin the the pages. 
The last patch adds two new udmabuf selftests to verify data coherency
after page migration.


This patch (of 3):

For drivers that would like to longterm-pin the pages associated with a
file, the pin_user_pages_fd() API provides an option to not only FOLL_PIN
the pages but also to check and migrate them if they reside in movable
zone or CMA block.  For now, this API can only work with files belonging
to shmem or hugetlbfs given that the udmabuf driver is the only user.

It must be noted that the pages associated with hugetlbfs files are
expected to be found in the page cache.  An error is returned if they are
not found.  However, shmem pages can be swapped in or allocated if they
are not present in the page cache.

Link: https://lkml.kernel.org/r/20231003074447.3245729-1-vivek.kasireddy@intel.com
Link: https://lkml.kernel.org/r/20231003074447.3245729-2-vivek.kasireddy@intel.com
Signed-off-by: Vivek Kasireddy <vivek.kasireddy@intel.com>
Suggested-by: Jason Gunthorpe <jgg@nvidia.com>
Cc: David Hildenbrand <david@redhat.com>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: Mike Kravetz <mike.kravetz@oracle.com>
Cc: Hugh Dickins <hughd@google.com>
Cc: Peter Xu <peterx@redhat.com>
Cc: Gerd Hoffmann <kraxel@redhat.com>
Cc: Dongwon Kim <dongwon.kim@intel.com>
Cc: Junxiao Chang <junxiao.chang@intel.com>
Cc: Shuah Khan <shuah@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>