mm/migrate: correct nr_failed in migrate_pages_sync()
In the upstream tree, large folios can be migrated via move_pages()
syscall and there is no VM_BUG_ON in do_move_pages_to_node() but pages
left in the source page list will not be returned to LRU lists.  This is
from my code inspection and I have not reproduced it on upstream tree yet.

nr_failed was missing the rc value from migrate_pages_batch() and can
cause a mismatch between migrate_pages() return value and the number of
not migrated pages, i.e., when the return value of migrate_pages() is 0,
there are still pages left in the from page list.  It will happen when a
non-PMD THP large folio fails to migrate due to -ENOMEM and is split
successfully but not all the split pages are not migrated,
migrate_pages_batch() would return non-zero, but astats.nr_thp_split =3D
0.  nr_failed would be 0 and returned to the caller of migrate_pages(),
but the not migrated pages are left in the from page list without being
added back to LRU lists.

Link: https://lkml.kernel.org/r/20231009203231.1715845-2-zi.yan@sent.com
Fixes: 2ef7dbb26990 ("migrate_pages: try migrate in batch asynchronously fi=
rstly")
Signed-off-by: Zi Yan <ziy@nvidia.com>
Cc: Baolin Wang <baolin.wang@linux.alibaba.com>
Cc: David Hildenbrand <david@redhat.com>
Cc: Huang Ying <ying.huang@intel.com>
Cc: Matthew Wilcox (Oracle) <willy@infradead.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>